gummiboot: Add Android boot image support
BZ: 138088

Additional file android.c contains logic to load and parse Android
boot images. Unlike the rest of GummiBoot it is BSD-licensed.
Some code was pulled in from EFILinux so that we can directly
start the Linux kernel. This makes it so we don't have to modify
the kernel EFI stub to load the ramdisk and can verify the entire
boot image all at once. We also add support for using the Android
BCB for recovery console boot persistence.

Any error loading an image stalls the bootloader for a few
seconds before returning to the menu. Security errors return to the
menu instead of exiting.

Change-Id: I74291546967929cfe26c3c38d7d21e71df408dc9
Signed-off-by: Andrew Boie <andrew.p.boie@intel.com>
Signed-off-by: Vincent Tinelli <vincent.tinelli@intel.com>
Reviewed-on: http://android.intel.com:8080/131794
Reviewed-by: Chouleur, Sylvain <sylvain.chouleur@intel.com>
Tested-by: Chouleur, Sylvain <sylvain.chouleur@intel.com>