improved Android command line handling
BZ: 138088

We don't want to have to sign the bootloader configs for Android.
So remove the ability to specify command line options there; it
was only being used to pass in the install id anyway.

Add a loader config option for install_id. If it exists, amend
the kernel command line with it. This is only used by the disk
scalabilty stuff and is optional.

Read the device's serial number out of SMBIOS and amend the kernel
command line with it. Android expects the serial number to be populated
by the bootloader.

Change-Id: I21e312f4b8e1ba281881e705ee8b15ffc3eddd8c
Depends-Change-Id: I41ee7affd604e84074267a61a6bec34d8499f13b
Signed-off-by: Andrew Boie <andrew.p.boie@intel.com>
Signed-off-by: Vincent Tinelli <vincent.tinelli@intel.com>
Reviewed-on: http://android.intel.com:8080/131798
Reviewed-by: Chouleur, Sylvain <sylvain.chouleur@intel.com>
Tested-by: Chouleur, Sylvain <sylvain.chouleur@intel.com>