android: Add capability to start boot image from file
BZ: 138088

This is desirable for the Iago installation media where
the boot image resides on the EFI system partition instead
of its own partition entry. We will then be able to enforce
security policy for the installation media.

-O0 wasn't supposed to be in the CFLAGS, remove it. Also
enforce -Werror since at the moment all this code compiles
without warnings.

Change-Id: I97c9bea1da35e548bbb6c44be9e88cb0041c12ba
Signed-off-by: Andrew Boie <andrew.p.boie@intel.com>
Signed-off-by: Vincent Tinelli <vincent.tinelli@intel.com>
Reviewed-on: http://android.intel.com:8080/131799
Reviewed-by: Chouleur, Sylvain <sylvain.chouleur@intel.com>
Tested-by: Chouleur, Sylvain <sylvain.chouleur@intel.com>