use shim StartImage, policy cleanups
BZ: 138088

() Policy for menu-less systems simplified to give the user
more control. Volume up is still a shortut to Recovery, but
any other key shows the menu. There are too many other useful
options to hard-code the only other working key to Fastboot,
especially on a dual boot or unlocked installation.

() If loading an image fails for any reason, simply print the
error and show the menu. The user can load the recovery console
if they want to, or use fastboot, enroll additional keys with
MokManager, etc. Don't assume that the only recourse when a
boot image fails to load is to instantly load the Recovery
Console; for example we might just need to load some keys, or
prefer to fix the problem with fastboot.

() Automatically add MokManager to the menu if we see it in the
EFI System Partition.

() Use new Shim API for starting EFI images if secure boot is
enabled. This will use the shim's security logic and additional
key management, needed to load EFI applications that aren't signed
with the BIOS DB key.

() Print a more informative message if secure boot is enabled
and user attempts to load an unsigned Android boot image.

Depends-Change-Id: Ie483ed009634116e288d5d15fed58e012073af6f
Change-Id: I930bd22c019a5bb4331b2d12b00360a408c15fea
Signed-off-by: Andrew Boie <andrew.p.boie@intel.com>
Signed-off-by: Vincent Tinelli <vincent.tinelli@intel.com>
Reviewed-on: http://android.intel.com:8080/131801
Reviewed-by: Chouleur, Sylvain <sylvain.chouleur@intel.com>
Tested-by: Chouleur, Sylvain <sylvain.chouleur@intel.com>