Implement OSNIB.
BZ: 143502

To manage multi-version of osnib, a generic structure osnib_gen is used
to expose all possible version's osnib data.

This osnib_gen is populated from the version specific osnib and when
bootlogic is finished, a hook operation is called which can
contain (this is the case for x86 platform) the write of OSNIB in nvram.

Change-Id: Icebe52b60661230bd7c18f84a1ea6d3970544b1e
Signed-off-by: Sylvain Chouleur <sylvain.chouleur@intel.com>
Reviewed-on: http://android.intel.com:8080/137679