target: use OsIndications EFI variable to switch to dnx mode
BZ: 154008

It has been decided that the dnx mode will not be supported as a load
option like it should have been. Finally, it relies on a dedicated and
absolutely out of the specifications path using the
OsIndicationsSupported and OsIndications EFI variables. The idea is,
in the context of the user asking for the "dnx" mode, we check for the
fifth bit availability in the OsIndicationsSupported bitmask EFI
variable and then we set the fifth bit in OsIndications bitmask EFI
variable accordingly. This fifth bit is named
EFI_OS_INDICATION_RESCUE_MODE and it indicates to the BIOS that we
request it to break the usual boot flow and go into Fastboot mode
instead.

This patch adds a new "command line" option -t <target> which is able
to boot registered partition targets as well as the dnx mode.

Change-Id: Ia8ac0eb078d8d262a572087a3bb2911358c935df
Signed-off-by: Jeremy Compostella <jeremy.compostella@intel.com>