EM: add variables to statically and dynamically choose the policy
BZ: 159307

In case of BIOS bug, hardware bug or during the PowerOn activity it is
very useful to be able to deactive some control like the EM controls
performed by the OSLoader.

This patch adds the use of the OSLOADER_EM_POLICY variable which
possible values are 'uefi' and 'fake'. 'uefi' is the policy based on
the data retrieved from UEFI while 'fake' is the emulation policy.

For engineering build only, the OSLoader provides a command line
option -e <policy> which allows to dynamically change the EM policy.

Change-Id: I6f224d3688858af3adde4f364a185a4a06f65fc2
Signed-off-by: Jeremy Compostella <jeremy.compostella@intel.com>