Trig a shuttdown upon battery insertion detected
BZ: 181704

In CrystalCove PMIC, battery insertion was not configured as a wake-capable
event in register SRCWAKECFG. hence we didnt see this issue.
In XPOWER, this is not configurable.
However, on earlier platforms, battery insertion is not a desired wake-event.

OSloader should trig a cold off if a battery insertion is the boot reason

Change-Id: Ie1f0f81ada9e7256a374702e19eb6615e84720cd
Signed-off-by: Benoit Fradin <benoit.fradin@intel.com>